feat: updated test
