+++ app/cdap/components/ConnectionList/Components/CustomTooltip/__test__/CustomTooltip.test.tsx
@@ -19,12 +19,14 @@ import { render, screen } from '@testing-library/react';
 import CustomTooltip from '../index';
 import { Typography } from '@material-ui/core';
 
-test('renders Connection list component', () => {
-  render(
-    <CustomTooltip arrow title={'Custom tooltip'}>
-      <Typography></Typography>
-    </CustomTooltip>
-  );
-  const ele = screen.getByTestId(/tooltip-parent/i);
-  expect(ele).toBeInTheDocument();
+describe('Test Connections List Component', () => {
+  it('Should render Connection list component', () => {
+    render(
+      <CustomTooltip arrow title={'Custom tooltip'}>
+        <Typography></Typography>
+      </CustomTooltip>
+    );
+    const ele = screen.getByTestId(/tooltip-parent/i);
+    expect(ele).toBeInTheDocument();
+  });
 });
