+++ app/cdap/components/ConnectionList/Components/TabLabelCanSample/mock/mockConnectorTypeData.ts
@@ -1,0 +1,7 @@
+export const mockConnectorTypeData = {
+  name: 'File',
+  type: 'connector',
+  category: 'File',
+  description: 'Connection to browse and sample data from the local file system.',
+  className: 'io.cdap.plugin.batch.connector.FileConnector',
+};
